Create logs-server.py, rename scraps dir to scripts
